* BugFix - Em alguns pontos um cast *int* era aplicado a position e size. Estas variáveis devem ser *long long*. O problema não aparecia pelo fato de se operar, nos testes feitos até o momento, com buffers de tamanho dentro do limite de *int* para máquinas 32bits.
git-svn-id: https://subversion.tecgraf.puc-rio.br/engdist/openbus/trunk@85006 ae0415b3-e90b-0410-900d-d0be9363c56b

+++ lib/ftc/include/ftc.h
@@ -32,8 +32,8 @@ class ftc {
     void setPosition( unsigned long long position ) ;
     unsigned long long getPosition() ;
     unsigned long long getSize() ;
-    void read( char* data, size_t nbytes, unsigned long long position ) ;
-    void write( char* data, size_t nbytes, unsigned long long position ) ;
+    void read( char* data, unsigned long long nbytes, unsigned long long position ) ;
+    void write( char* data, unsigned long long nbytes, unsigned long long position ) ;
 } ;
 
 #endif
